New indexed render states management.
Till now indexed render states such as lights, clipping planes and texture unit related render states had their binding index hardwired in the render state class making it impossible to share the same indexed render states among different shaders with a different index. Now it is possible for a light, a clipping plane etc. to be bound to the same or a different shader with the same or different binding point, maximizing render states reusability and simplifying the code necessary to their management.