IMPORTANT: Matrix4::getLookAt() changed semantic, now returns the view-look-at-matrix rather than the modeling-look-at-matrix.
For better compliance with OpenGL and COLLADA conventions Matrix4::getLookAt() now returns a matrix that is suitable to be used as a camera view matrix instead of returning a modeling matrix suitable to be used as a local matrix in Transform as before. 
The new method Matrix4::getLookAtModeling() returns the modeling look-at matrix that was previously returned by getLookAt().
The same is true for getAsLookAt() and getAsLookAtModeling().